+++ nuxeo-drive-client/nxdrive/direct_edit.py
@@ -101,12 +101,8 @@ class DirectEdit(Worker):
             info = parse_protocol_url(unicode(url))
         if info is None:
             return
-        # Handle backward compatibility
-        if info.get('item_id') is not None:
-            self.edit(info['server_url'], info['item_id'])
-        else:
-            self.edit(info['server_url'], info['doc_id'], user=info['user'],
-                      download_url=info['download_url'])
+        self.edit(info['server_url'], info['doc_id'], user=info['user'],
+                  download_url=info['download_url'])
 
     def _cleanup(self):
         log.debug("Cleanup DirectEdit folder")
@@ -295,22 +291,14 @@ class DirectEdit(Worker):
 
     def edit(self, server_url, doc_id, user=None, download_url=None):
         try:
-            log.debug("Editing doc %s on %s", doc_id, server_url)
-            # Handle backward compatibility
-            if '#' in doc_id:
-                engine = self._get_engine(server_url)
-                if engine is None:
-                    log.warning(
-                        'No engine found for %s, cannot edit file with remote ref %s',
-                        server_url, doc_id)
-                    return
-                self._manager.edit(engine, doc_id)
-            else:
-                # Download file
-                file_path = self._prepare_edit(server_url, doc_id, user=user, download_url=download_url)
-                # Launch it
-                if file_path is not None:
-                    self._manager.open_local_file(file_path)
+            log.debug('Editing doc %s on %s', doc_id, server_url)
+            # Download file
+            file_path = self._prepare_edit(
+                server_url, doc_id, user=user, download_url=download_url)
+
+            # Launch it
+            if file_path is not None:
+                self._manager.open_local_file(file_path)
         except OSError as e:
             if e.errno == 13:
                 # open file anyway